CanYouAccess Java Hackerrank Solution
You are given a class Solution and an inner class Inner. Private. The main method of class Solution takes an integer num as input. The powerof2 in class Inner.Private checks whether a number is a power of. You have to call the method powerof2 of the class Inner. Private from the main method of the class Solution.